feat: sistema completo de reservas, notificaciones, facturas y UI profesional
- feat(backend): implementar notificaciones con observer y políticas
- feat(backend): corregir buffer de limpieza en CalendarService
- feat(backend): añadir endpoints de documentos por usuario
- feat(backend): optimizar generación de facturas con logo comprimido
- feat(frontend): crear componente back-button reutilizable
- feat(frontend): implementar paginación en notificaciones
- feat(frontend): conectar documentos al backend real
- feat(frontend): añadir modal para términos y condiciones
- fix(frontend): eliminar alerts y usar modales profesionales
- fix(backend): corregir error de memoria en PdfService
- fix(calendar): liberar fechas al cancelar reserva
- refactor(frontend): centralizar utilidades de fechas y estados
- docs(README): actualizar documentación del proyecto

+++ backend/database/seeders/DocumentSeeder.php
@@ -58,7 +58,7 @@ class DocumentSeeder extends Seeder
         
         // Documentos para guests (30%)
         $guests = Guest::all();
-        foreach ($guests->random($guests->count() * 0.3) as $guest) {
+        foreach ($guests->random((int) ($guests->count() * 0.3)) as $guest) {
             Document::create([
                 'documentable_type' => Guest::class,
                 'documentable_id' => $guest->id,
